requirements.txt
# Core Django and web framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Django Channels for WebSocket support (Real-time communication)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Async task processing
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Database
psycopg2-binary==2.9.9

# Data science dependencies
pandas==2.1.3
numpy==1.26.2

# Configuration and utilities
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
uvicorn[standard]==0.24.0
django-filter==23.3

# HTTP client for external API calls (Spotify, MusicBrainz)
requests==2.31.0

# OpenAPI schema and interactive docs
drf-spectacular==0.27.0

# Celery monitoring
flower==2.0.1

# Load testing (development only)
# locust==2.20.0

# Security scanning (development only)
# pip-audit
# bandit
